Remove Dependency on package.json "build" Script
Previously, there was a step that ran "npm run build", but this assumes
that the consumer has a "build" script in their package.json. While this
is usually a same assumption since many people start off using
create-react-app which initializes the package.json with the "build"
script, some consumers may change the name of that script.

With this commit, the step is no longer running an npm script but,
instead, uses npx to run the react-scripts package directly to invoke
the react-script build script. With this change, consumers do not need
to keep the build script configured in their package.json. Of course,
even if they do continue to have that script, this change will still
work for those consumers passively.